+++ client/inc/System/Shoot/ShootSystem.hh
@@ -35,7 +35,7 @@ namespace Pattern {
 		if (s == Side::LEFT)
 			vel *= -1;
 		pos.first += vel;
-		pos.second += (35 * sin(pos.first * 0.5 * M_PI / 180));
+		pos.second += static_cast<float>(35 * sin(pos.first * 0.5 * M_PI / 180));
 		e.manager.set("position", pos);
 	}
 }
